% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{k_analysis_columns}
\alias{k_analysis_columns}
\title{Preprocess Text Features}
\usage{
k_analysis_columns(n = 500)
}
\arguments{
\item{n}{Number of columns or components.}
}
\value{
Preprocessed data frame.
}
\description{
Cleans and prepares raw feature vectors for analysis.
}
