% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{k_read_features}
\alias{k_read_features}
\title{Read Features CSV}
\usage{
k_read_features(path, text_id_col = "text_id")
}
\arguments{
\item{path}{Path to feature CSV}

\item{text_id_col}{Name of the column containing text identifiers.}
}
\value{
Data frame
}
\description{
Read Features CSV
}
